+++ packages/moonstone/Input/InputSpotlightDecorator.js
@@ -27,6 +27,24 @@ const handleKeyDown = handle(
 	call('onKeyDown')
 );
 
+
+/**
+ * Default config for [InputSpotlightDecorator]{@link moonstone/Input.InputSpotlightDecorator}
+ *
+ * @memberof moonstone/Input/InputSpotlightDecorator.InputSpotlightDecorator
+ * @hocconfig
+ */
+const defaultConfig = {
+	/**
+	 * Suppress the pointer lock behavior of moonstone input
+	 *
+	 * @type {Boolean}
+	 * @default false
+	 * @memberof moonstone/Input/InputSpotlightDecorator.InputSpotlightDecorator.defaultConfig
+	*/
+	noLockPointer: false
+};
+
 /**
  * A higher-order component that manages the
  * spotlight behavior for an {@link moonstone/Input.Input}
@@ -36,7 +54,8 @@ const handleKeyDown = handle(
  * @hoc
  * @private
  */
-const InputSpotlightDecorator = hoc((config, Wrapped) => {
+const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
+	const {noLockPointer} = config;
 	const Component = Spottable({emulateMouse: false}, Wrapped);
 	const forwardBlur = forward('onBlur');
 	const forwardMouseDown = forward('onMouseDown');
@@ -137,7 +156,9 @@ const InputSpotlightDecorator = hoc((config, Wrapped) => {
 					onSpotlightDisappear();
 				}
 
-				releasePointer(this.state.node);
+				if (!noLockPointer) {
+					releasePointer(this.state.node);
+				}
 			}
 		}
 
@@ -155,11 +176,15 @@ const InputSpotlightDecorator = hoc((config, Wrapped) => {
 			if (focusChanged) {
 				if (this.state.focused === 'input') {
 					forward('onActivate', {type: 'onActivate'}, this.props);
-					lockPointer(this.state.node);
+					if (!noLockPointer) {
+						lockPointer(this.state.node);
+					}
 					this.paused.pause();
 				} else if (prevState.focused === 'input') {
 					forward('onDeactivate', {type: 'onDeactivate'}, this.props);
-					releasePointer(prevState.node);
+					if (!noLockPointer) {
+						releasePointer(prevState.node);
+					}
 					this.paused.resume();
 				}
 			}