+++ packages/moonstone/Spinner/Spinner.js
@@ -92,7 +92,7 @@ const SpinnerBase = kind({
 	propTypes: /** @lends moonstone/Spinner.SpinnerBase.prototype */ {
 		/**
 		 * Customizes the component by mapping the supplied collection of CSS class names to the
-		 * corresponding internal Elements and states of this component.
+		 * corresponding internal elements and states of this component.
 		 *
 		 * The following classes are supported:
 		 *